Integrated Movies list API
