+++ src/index.js
@@ -1,15 +1,18 @@
-import 'bootstrap/dist/css/bootstrap.css';
-import _ from 'lodash';
-import './style.css';
+import "bootstrap/dist/css/bootstrap.css";
+import "./style.css";
+import { getSchedule } from "./schedule";
 
-function component() {
-  const element = document.createElement('div');
+const displaySchedules = () => {
+  const schedules = getSchedule();
+  //const table = document.getElementById("table-scores");
+  schedules.then((schedule) => {
+    schedule.forEach((item) => {
+      let info = item._embedded;
+      console.log(info.show.name);
+      console.log(info.show.id);
+    });
+    //table.innerHTML = tableScoreRow;
+  });
+};
 
-  // Lodash, now imported by this script
-  element.innerHTML = _.join(['Hello', 'webpack'], ' ');
-  element.classList.add('hello');
-
-  return element;
-}
-
-document.body.appendChild(component());
+displaySchedules();
