+++ webpack.config.js
@@ -3,6 +3,7 @@ const HtmlWebpackPlugin = require("html-webpack-plugin");
 
 module.exports = {
   entry: "./src/index.js",
+  mode: "development",
   devServer: {
     static: "./dist",
   },
